requirements.txt
PyQt5>=5.15.0
pillow>=8.0.0
qrcode>=7.3.1
python-dateutil>=2.8.2
Flask>=2.0.0
Flask-Bcrypt>=1.0.1 